--- src/main/java/com/fourthread/ozang/module/domain/security/exception/SecurityException.java
@@ -1,12 +1,0 @@
-package com.fourthread.ozang.module.domain.security.exception;
-
-import com.fourthread.ozang.module.common.exception.ErrorDetails;
-import com.fourthread.ozang.module.common.exception.GlobalException;
-
-public class SecurityException extends GlobalException {
-
-  public SecurityException(String exceptionName, String message,
-      ErrorDetails details) {
-    super(exceptionName, message, details);
-  }
-}
