+++ src/main/java/com/fourthread/ozang/module/domain/security/filter/JsonLoginFilter.java
@@ -1,15 +1,11 @@
 package com.fourthread.ozang.module.domain.security.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fourthread.ozang.module.common.exception.ErrorCode;
-import com.fourthread.ozang.module.domain.security.SecurityMatchers;
+import com.fourthread.ozang.module.domain.security.jwt.dto.type.SecurityMatchers;
 import com.fourthread.ozang.module.domain.user.dto.request.LoginRequest;
-import com.fourthread.ozang.module.domain.user.entity.User;
-import com.fourthread.ozang.module.domain.user.exception.UserException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpMethod;
