+++ src/main/java/com/fourthread/ozang/module/domain/security/handler/JwtLogoutHandler.java
@@ -1,5 +1,6 @@
-package com.fourthread.ozang.module.domain.security.jwt;
+package com.fourthread.ozang.module.domain.security.handler;
 
+import com.fourthread.ozang.module.domain.security.jwt.JwtService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
