+++ src/main/java/com/fourthread/ozang/module/domain/security/oauth/parser/OAuthAttributesParser.java
@@ -1,6 +1,6 @@
-package com.fourthread.ozang.module.domain.oauth.parser;
+package com.fourthread.ozang.module.domain.security.oauth.parser;
 
-import com.fourthread.ozang.module.domain.oauth.dto.OAuthAttributes;
+import com.fourthread.ozang.module.domain.security.oauth.dto.OAuthAttributes;
 import java.util.Map;
 
 public interface OAuthAttributesParser {
